+++ R/compat-vctrs-helpers.R
@@ -16,6 +16,7 @@ new_tune_results_from_template <- function(x, to) {
     x = x,
     parameters = attrs$parameters,
     metrics = attrs$metrics,
+    outcomes = attrs$outcomes,
     rset_info = attrs$rset_info
   )
 }
@@ -104,6 +105,7 @@ new_resample_results_from_template <- function(x, to) {
     x = x,
     parameters = attrs$parameters,
     metrics = attrs$metrics,
+    outcomes = attrs$outcomes,
     rset_info = attrs$rset_info
   )
 }
@@ -132,6 +134,7 @@ new_iteration_results_from_template <- function(x, to) {
     x = x,
     parameters = attrs$parameters,
     metrics = attrs$metrics,
+    outcomes = attrs$outcomes,
     rset_info = attrs$rset_info
   )
 }