+++ R/resample.R
@@ -159,6 +159,7 @@ resample_workflow <- function(workflow, resamples, metrics, control) {
     x = resamples,
     parameters = parameters(workflow),
     metrics = metrics,
+    outcomes = outcome_names(workflow),
     rset_info = rset_info
   )
 }